requirements.txt
grpcio
grpcio-tools
redis
pika
termcolor
prompt_toolkit